handling buttons a little differently. this will fail tests
